+++ integrations/clickup/clickup/client.py
@@ -2,18 +2,29 @@ from typing import Any, AsyncGenerator
 import httpx
 from loguru import logger
 from port_ocean.utils import http_async_client
-
+from port_ocean.context.ocean import ocean
 
 
 class ClickUpClient:
     def __init__(self, clickup_url: str, clickup_api_key: str) -> None:
         self.clickup_url = clickup_url
         self.api_url = f"{self.clickup_url}/api/v2"
-
         self.api_auth_header = {"Authorization": clickup_api_key}
         self.client = http_async_client
         self.client.headers.update(self.api_auth_header)
         self.client.timeout = httpx.Timeout(30)  # 30 seconds timeout for requests
+
+    @staticmethod
+    def get_clickup_client():
+        try:
+            host = ocean.integration_config.get("clickup_host")
+            api_key = ocean.integration_config.get("clickup_api_key")
+            if not host or not api_key:
+                raise ValueError("ClickUp host or API key is not configured properly.")
+            return ClickUpClient(host, api_key)
+        except KeyError as e:
+            logger.error(f"Configuration key missing: {e}")
+            raise
 
     @staticmethod
     def parse_projects_from_folders(folders: list[dict[str, Any]]) -> list[dict[str, Any]]:
@@ -46,11 +57,9 @@ class ClickUpClient:
     async def _get_folders(self, space_id: str) -> dict[str, Any]:
         return await self._get(f"/space/{space_id}/folder")
 
-    async def get_projects(
-        self, space_id: str
-    ) -> AsyncGenerator[list[dict[str, Any]], None]:
+    async def get_projects(self, space_id: str, team_id: str) -> list[dict[str, Any]]:
         logger.info("Getting projects from ClickUp")
-        
+
         try:
             foldered_projects = self.parse_projects_from_folders(
                 (await self._get_folders(space_id))["folders"]
@@ -61,32 +70,40 @@ class ClickUpClient:
             if not projects:
                 logger.warning("Space query returned 0 projects")
 
-            yield projects
+            return [{**project, "__team_id": team_id} for project in projects]
         except httpx.HTTPStatusError as e:
-            logger.error(f"Failed to fetch projects for space_id {space_id} : {e}")
-            yield []
+            logger.error(f"Failed to fetch projects for space_id {space_id}: {e}")
+            return []
 
-    async def get_teams(self) -> AsyncGenerator[list[dict[str, Any]], None]:
+    async def fetch_all_projects(self):
+        teams = await self.get_teams()
+        for team in teams:
+            team_id = team["id"]
+            spaces = await self.get_spaces(team_id)
+            logger.info(f"Received spaces batch with {len(spaces)} spaces for team {team_id}")
+            for space in spaces:
+                space_id = space["id"]
+                projects = await self.get_projects(space_id, team_id)
+                return projects, team_id
+
+    async def get_teams(self) -> list[dict[str, Any]]:
         logger.info("Getting teams from ClickUp")
         try:
             total_teams = (await self._get_teams())["teams"]
-            yield total_teams
+            return total_teams
         except httpx.HTTPStatusError as e:
-            logger.error(f"Failed to fetch teams : {e}")
-            yield []
+            logger.error(f"Failed to fetch teams: {e}")
+            return []
 
-    async def get_spaces(
-        self, team_id: str
-    ) -> AsyncGenerator[list[dict[str, Any]], None]:
+    async def get_spaces(self, team_id: str) -> list[dict[str, Any]]:
         logger.info("Getting spaces from ClickUp")
 
         try:
             space_response_list = (await self._get_spaces(team_id))["spaces"]
-            yield space_response_list
+            return space_response_list
         except httpx.HTTPStatusError as e:
             logger.error(f"Failed to fetch spaces for team_id: {team_id}: {e}")
-            yield []
-
+            return []
 
     async def get_paginated_tasks(
         self, list_id: str
@@ -95,7 +112,7 @@ class ClickUpClient:
 
         params = self._generate_base_req_params()
 
-        page = params.get("page",0)
+        page = params.get("page", 0)
         while True:
             logger.info(f"Current query position: page {page}")
             try:
@@ -109,3 +126,15 @@ class ClickUpClient:
                 break
             page += 1
             params["page"] = page
+
+    async def fetch_issues(self):
+        projects, _ = await self.fetch_all_projects()
+        for project in projects:
+            project_id = project["id"]
+            try:
+                async for tasks in self.get_paginated_tasks(project_id):
+                    logger.info(f"Received task batch with {len(tasks)} tasks for project {project_id}")
+                    yield tasks
+            except Exception as e:
+                logger.error(f"Failed to fetch tasks for project {project_id}: {e}")
+                raise
